code/requirements.txt
actionlib==1.14.0
amqp==5.0.9
angles==1.9.13
aniso8601==9.0.1
async-timeout==4.0.3
billiard==3.6.4.0
blinker==1.5
bondpy==1.8.6
cachelib==0.9.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
celery==5.2.3
cffi==1.17.1
chardet==5.2.0
click==8.1.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
contourpy==1.2.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==3.4.8
cv-bridge==1.16.2
cycler==0.12.1
dbus-python==1.2.18
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.7.0
distro-info==1.0
dnspython==2.3.0
dynamic-reconfigure==1.7.3
email-validator==1.3.1
ephem==4.1.3
Flask==2.2.3
Flask-Caching==2.1.0
Flask-Excel==0.0.7
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.46.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
greenlet==2.0.2
httplib2==0.20.2
idna==3.4
image-geometry==1.16.2
importlib-metadata==6.0.0
interactive-markers==1.12.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keyring==23.5.0
kiwisolver==1.4.5
kombu==5.2.3
laser_geometry==1.6.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lml==0.1.0
MarkupSafe==2.1.2
matplotlib==3.10.0
message-filters==1.17.0
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
packaging==23.2
passlib==1.7.4
Pillow==10.1.0
prompt-toolkit==3.0.28
pycparser==2.22
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-memcached==1.58
python-qt-binding==0.4.4
pytz==2023.3.post1
PyYAML==6.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
redis==5.0.1
resource_retriever==1.12.7
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.25
SecretStorage==3.3.1
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
SQLAlchemy==2.0.6
systemd-python==234
texttable==1.7.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
topic-tools==1.17.0
typing_extensions==4.5.0
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.5
Werkzeug==2.2.3
WTForms==3.0.1
xacro==1.14.18
zipp==3.15.0